Moved RewardAndPenalty to its own class
