+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/statetransition/epoch/RewardAndPenaltyDeltas.java
@@ -16,8 +16,6 @@ package tech.pegasys.teku.spec.logic.common.statetransition.epoch;
 import com.google.common.base.MoreObjects;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
 public class RewardAndPenaltyDeltas {
   private final List<RewardAndPenalty> deltas;
@@ -36,56 +34,5 @@ public class RewardAndPenaltyDeltas {
   @Override
   public String toString() {
     return MoreObjects.toStringHelper(this).add("deltas", deltas).toString();
-  }
-
-  public static class RewardAndPenalty {
-    private UInt64 reward = UInt64.ZERO;
-    private UInt64 penalty = UInt64.ZERO;
-
-    public void reward(final UInt64 amount) {
-      reward = reward.plus(amount);
-    }
-
-    public void penalize(final UInt64 amount) {
-      penalty = penalty.plus(amount);
-    }
-
-    public void add(final RewardAndPenalty other) {
-      reward(other.reward);
-      penalize(other.penalty);
-    }
-
-    public UInt64 getReward() {
-      return reward;
-    }
-
-    public UInt64 getPenalty() {
-      return penalty;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-      if (this == o) {
-        return true;
-      }
-      if (o == null || getClass() != o.getClass()) {
-        return false;
-      }
-      final RewardAndPenalty delta = (RewardAndPenalty) o;
-      return Objects.equals(reward, delta.reward) && Objects.equals(penalty, delta.penalty);
-    }
-
-    @Override
-    public int hashCode() {
-      return Objects.hash(reward, penalty);
-    }
-
-    @Override
-    public String toString() {
-      return MoreObjects.toStringHelper(this)
-          .add("reward", reward)
-          .add("penalty", penalty)
-          .toString();
-    }
   }
 }
