+++ ethereum/spec/src/testFixtures/java/tech/pegasys/teku/spec/util/DataStructureUtil.java
@@ -183,6 +183,7 @@ import tech.pegasys.teku.spec.datastructures.type.SszSignature;
 import tech.pegasys.teku.spec.datastructures.util.DepositGenerator;
 import tech.pegasys.teku.spec.executionlayer.ForkChoiceState;
 import tech.pegasys.teku.spec.executionlayer.PayloadBuildingAttributes;
+import tech.pegasys.teku.spec.logic.common.statetransition.epoch.RewardAndPenalty;
 import tech.pegasys.teku.spec.logic.common.statetransition.epoch.RewardAndPenaltyDeltas;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitions;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsAltair;
@@ -2413,8 +2414,7 @@ public final class DataStructureUtil {
     final RewardAndPenaltyDeltas rewardAndPenaltyDeltas =
         new RewardAndPenaltyDeltas(validatorCount);
     for (int i = 0; i < validatorCount; i++) {
-      final RewardAndPenaltyDeltas.RewardAndPenalty rewardAndPenalty =
-          rewardAndPenaltyDeltas.getDelta(i);
+      final RewardAndPenalty rewardAndPenalty = rewardAndPenaltyDeltas.getDelta(i);
       rewardAndPenalty.reward(randomUInt64(1000));
       rewardAndPenalty.penalize(randomUInt64(1000));
     }
